Describe Peer's real marketplace flow; share the wallet-address hint
Peer's docs describe a Buy tab that no longer exists — the trade page
is a marketplace of liquidity-provider offers. Steps now cover picking
currency + payment app, choosing an offer, and automatic payment
verification. Extracts WalletAddressHint so both provider panels show
the connected address with a copy button.
